Use a more robust mechanism to extract the tag start/end indices
As of now in order to determine the range of the points in Mhistory that
correspond to a specific module we were searching for an exact match of
the tag (TAG_s or TAG_e). However this precludes the use of any other
information in the tags (TAG_smth_e) that is useful for most modules.
Now instead of an exact match the tags are searched only with startswith
(for the TAG part)and endswith (for the _s, _i and _e part).