Add support for gzipped session files
Session files tend to become very large, for example with long endurance
or multibit testing, and they tend to creep in the 10s of MB. As the
session files are text they compress well with gzip so the option to
save (or load) csv.gz files has been added. The modification is very
simple: if the target file is csv.gz instead of csv the open function is
replaced with gzip.open. Since the semantics of both open and gzip.open
are the same they work identically as far as the application is concerned.