Properly configure ArC ONE for TIA4P read when Vread < 0
When TIA4P mode is setup for reading the read option sent to ArC ONE is
always 2. However to properly read with negative voltages the option 3
must be sent because otherwise the microcontroller returns garbage data.
So, when Vread is negative and the global read option is 2 (TIA4P)
properly communicate the read configuration to ArC ONE by sending code 3
instead of 2.