Support linear progression (fixed step) of pulse widths in FormFinder.
Until now the pulse width progression in FormFinder was only
multiplicative. For example for an 100% step and 1 us initial pulse width
the applied pulses would be 1 us, 2 us, 4 us, 8 us, etc. This is useful
when electroforming because in general similar pulse widths will have
similar results so one needs to scan many orders of magnitude quickly.
However, depending on application, one might need to have linear evolution
of the pulses with a fixed pulse width step. This option has now been
added to the interface with a new combobox.

Obviously this required firmware work to support it. As such FormFinder was
slightly edited to support an option for either multiplicative or linear
progression. To maintain some degree of backwards compatibility it was
assigned a new job number (141) and the old FormFinder (job 14) is now a
special case of the new general firmware routine. This assures the
compatibility from the firmware to older interfaces. However in order for
the interface to also support older firmwares the job number used for the
multiplicative case has remained as 14. So if the interface is used with an
older firmware it will still work for the multiplicative case. Obviously
the linear will not be supported.