Add a loading indication in the command line when opening files
As loading files can take long a progress indication has been added to
the command line. This will be upgraded into a regular progress bar in
the future. To calculate the byte size the fields of the csv file are
encoded into bytes and counted. Adding the size of the line terminators
per line and the commas should be accurate enough providing there are
no spaces between the csv fields.