Make sure that R{max,min}SET are properly populated as 1-D arrays
optimize.curve_fit always expects an 1-d array for the initialisation. So
far R{max,min}SET were initialised as a series of results from previous
curve fittings. However curve_fit returns a complicated mix of results
but we only need the first result (the actual fitted parameter). So we
might as well pre-allocate the R{max,min}SET arrays and populate it
accordingly. This way we ensure that both will be 1-d float64 arrays.