Make `Graphics` into a proper module; load data with `importlib`
Instead of loading graphics using absolute path `Graphics` is now a
proper module. Pixmaps are loaded with the `initialise` method and are
presented to the application itself through the `getIcon` or `getPixmap`
methods using `importlib`. This has the added benefit of caching the
pixmaps instead of loading them again every time they are requested.