Move common threading functionality to a base utility class
A lot of instrument communication is done with threads and changes are
communicated to the interface through signals. There was quite a lot of
duplicate code among different modules. For this reason a utility class
has been created that wraps the basic signals of these threads as well
as provides a simple decorator to toggle the interface depending on the
status of the active thread. This does not change the existing semantics
but reduces code by a lot.