Fix tag handling mechanism when populating the history tree
There are two "classes" of tags used throughout the program. The pulse
and read tags ('P', 'S R' and 'F R') that are handled separately and the
underscore-delimited tags that cover most of the modules. This commit
slightly alters the tag handling to look first for the pulse/read tags
and then fall back into the '_'-delimited case. As intermediate tags are
not used for marking the start/end of tests they can be ignored. This
also covers the corner case of STDP where the intermediate tags do not
have the usual '_i' suffix. Since they have no functional role in
marking the tests they can be safely ignored.

Handling of the tags is probably not optimal. The same can be said about
how they are abused to delimit tests but changing this behaviour now
would lead to many broken session logfiles.