Merge updateHistory functionality into a single function
Instead of using the `updateHistory` and `updateHistory_CT` functions
to update data the `updateHistory` function is accepting an additional
argument that defaults to `Vread`. As such CurveTracer (which has a
varying `Vread`) now overrides the `sendData` signal to take advantage
of the full form `updateHistory` without needing a special function.