Initial pypification of the project
This is the first commit in a series of efforts to pypi-fy the
arc1_pyqt structuring it as a proper Python project. All modules have
now been moved under `arc1pyqt`. Several modifications have been made
across different modules to ensure that the import order is correct.
Since now the project is namespaced `modutils` has been moved to the
root of the module directory.

Additionally a `setup.py` script has been added that takes care of
assembling the project and generating interface files (as a result
`genui.py` has been removed).