Consolidate tag and callback dicts and simplify module loading
This commit consolidates all module loading mechanisms used throughout
the program with a single operation. During startup all valid module
directories are scanned for compatible modules and the `modules`
dictionary is populated accordingly. Using a global variable is
definitely not optimal but now there is at least a single point of origin
for all modules. This commit will also break module loading as it exists
now but required alterations have been kept at minimum.  In order for a
file to be a valid module it requires a) a class subclassing
`BaseProgPanel` named the same as the filename and b) the presence of a
special variable called `tags`. This variable is a dict defining how the
toplevel module should be named, its short unique identifier as well as a
display function should this be required. This is a new namedtuple called
`ModTag`. There should be at least one `ModTag` present under the `top`
key of `tags` dict. If additional tags are needed then they can be
provided as a list under the `subtags` key of the `tags` dict. That being
said subtags should be used only when absolutely necessary to keep code
simple.

Finally this commit allows loading of modules from system and user
directories instead of the `ProgPanels` folder since `ProgPanels` should
absolutely be read-only (especially in a pypified installation).
Additional modules can be loaded from the `AppData` directory of the
specific platform as specified by `QtCore.QStandardPaths`. Local (user)
directories take precedence over global.  On Linux this user directory
would be `$HOME/.local/share/arc1pyqt/ProgPanels`, on Windows it would be
`%APPDATA%\arc1pyqt`, etc. Of course for more system-wide configurations
the expected system-wide dirs also apply: `/usr/share/arc1pyqt` on Linux,
`%PROGRAMDATA%\arc1pyqt` on Windows and so on.

This is part of an effort to make `arc1pyqt.modutils` the singular module
of reference for plugin loading and development rather than have
scattered bits and pieces around the whole codebase.