Tidy up state management across arc1_pyqt
arc1_pyqt relies quite significantly in global variables. This is a
not very efficient design and it's error prone but because of legacy
reasons making arc1_pyqt completely stateless practically requires a
complete software redesign. In an effort to mitigate this problem this
commit tidies up access to global variables and tries to abstract away
any information that is not required for the regular lifecycle of the
tool.

Any variables that are read only are moved into a separate constants
module. Everything else is moved to the new `state` module. This
module contains three dataclasses that describe the status of three
different aspects of the program: GUI application (`Application`) for
data presentation, program panels and syncronisation; crossbar
status (`Crossbar`) for current status of the devices across all
crosspoints and hardware (`Hardware`) for configuration and access to
ArC1 (or VirtualArC). Consumer modules can import `state` and then
access the relevant parts. The design is by no means ideal but is more
intuitive that the original design and in almost all cases mutating
is done from a single point at a time.

Instrument access is wrapped around an `Instrument` class which
presently can be either the physical ArC1 or VirtualArC. The frontend is
completely oblivious to that and the only point of access is
`state.hardware`. Instead of directly accessing the serial port
everything is done through `Instrument` implementers that also safely
expose some of the `Serial` functions whenver they are required.