Add an svg loading mechanism for `Graphics`
This adds preliminary support for SVG graphics loading. SVGs are loaded
into a separate array in `Graphics` and can be accessed with two new
functions `getSvgRenderer` and `getSvgScaled`. The first returns the
underlying `QSvgRenderer` so that it can be used directly with a
`QPainter` to paint on a widget. The latter returns the SVG rendered as
a scaled (optionally) pixmap to display on screen.