Add a logo display widget
A common occurrence across the codebase is the presentation of a large
coloured empty label with an image displayed on top. In order to
facilitate some re-usability this commit adds a new widget that exposes
this kind of functionality by displaying pixmaps or SVGs atop a coloured
background at different alignments.