Add wrappers for pyuic and pyrcc in modutils
When designing modules it's not unusual to use Qt Designer for the
UI definition. In order to allow modules to generate python classes
of UI files two additional functions have been added to modutils. The
first generates python files out of UI definitions (compile_ui) and
the second (compile_rc) bundles resources (qrc files) into a
loadable python module. If the UI makes use of resources the qrc files
MUST be processed first in order for them to be accessible to the
generated python definitions.