Ensure that Pulses and Reads are marked as such in history
When traversing the tag keys for read/pulse special tags only matching
against existence was done. So during this procedure the tag key
matched was always 'S R' or 'Read' (the first in the traversal list).
So all pulses and reads were marked as 'Read' in history, which is
obviously wrong. Now an additional matching between `tagString` (the
actual tag) and `tagKey` (the tag we are currently traversing) is done
to fix this.