Rework history tree panel
History tree panel was based on QTreeWidget that uses QStandardItems
for its nodes. Although this worked it required a lot of hacks to
include all the information that was relevant for each node. This has
now been reworked to use a QTreeView with a custom model that is
suitable for the data. This eliminated lots of duplicate code as well
as loads of unnecessary loops, such as recreating the list of tags on
history for the given entry FOR EVERY entry that's under processing.
This alone resulted in a ~20% speedup when opening large session files.
This approach also addresses a long standing issue where the order
of child entries under a toplevel W,B combination in the tree is
reversed when running on live devices (newest first) versus when being
loaded from file (oldest first). Now it's always oldest first. Whether
that's better or not remains to be seen but it is, at least, consistent.