Don't traverse history backwards when adding new history items
This commit addresses a long standing issue when building the history
tree from input data (whether live or loaded from file). Previously
every time an end tag is encountered (`_e`) the history was being
traversed in reverse from the point of insertion until the
corresponding start tag (`_s`) is found. This was very wasteful and
data files with many short-lived experiments would take a very long
time to parse. With this commit every time an `_s` tag is inserted the
start index is also stored in the history. Whenever an end tag is
subsequently found the index of the `_s` tag is popped and added to
the last column of `CB.history[w][b]`. So the range of that specific
module can be quickly retrieved without resorting to reverse traversing
the history every time an end tag is encountered. Since these indices
are ephemeral they are only available in the in-memory history and
are not saved in the session file. The addition of the start index
is done transparently to the modules and should not affect user
modules as long as they emit `sendData`s properly. Only exception is
CT_LIVE that does history manipulation.

This approach also handles corrupted data (i.e. incomplete blocks of
data) a bit more gracefully by altering unclosed blocks on the fly.
This is done when more than one `_s` tags are currently 'active'. In
that case we search for the last found non-final tag (typically of the
sort `XXX_i` or `XXX_i_Y`) and this one is replaced with an end tag.
This allows for the corrupted block to be delimited properly. Whether
the recovered data actual make any sense is up to the user to determine.
Only problem with this workaround is that we can only know if a
corrupted block was found only after we cross into the next one (since
the history tree is refreshed only upon encountering a final tag,
typically `XXX_e` or a read/pulse one - `S RX`, `F RX`, `P`). In this
particular case the history entry for that specific problematic word-
and bitline combination must be rebuilt. A new signal has been created
for that in the previous commit and it's emitted by the `addStartTag`
function of `state.Crossbar` when this situation occurs.