Add functions to associate panel widgets with a key in modutils
In order to persist data for programming panels (for SuperMode for
instance) the parameters of a panel need to be extracted somehow. This
commit adds two functions to either extract data or load data from/onto
the panel. Functions `{extract,set}PanelData` will automatically
assign values to QLineEdits, QComboBoxes, QSpinBoxes and
QDoubleSpinBoxes as long as they have been registered for automatic
extraction with `registerPropertyWidget`. This essentially assigns a
Qt property named `key` with a user-defined value, for instance
`self.registerPropertyWidget(anEdit, 'repetitions')`. These values can
be then applied to the panel's widgets with `setPanelData`. So, by
registering the widgets the super-class will provide automatic parameter
extraction and setup from/to the panel. For data that are more complex
than values in these four widgets the child class can override these
two functions to provide additional functionality.