Update SuperMode and modules to use the registered widget facility
This commit updates all modules to use the `registerPropertyWidget`
introduced in the previous commit. The save/load functionality of
SuperMode now takes advantage of this mechanism to produce text-only
json measurement chains as output files instead of the binary pickles
used before. Pickled objects work but they are heavily dependent on the
Python version used as well as the internal structure of the widgets.
For most programming panels the only change required was removing the
old `extractPanelParameters` and `setPanelParameters` functions and
instead register the relevant widgets with `registerPropertyWidget`.
This functionality is now provided automatically by the new
`{set,extract}PanelData` functions of `BaseProgPanel`. Only exception
to this is STDP which still requires some further data to be saved
along with the values of its child widgets.

The new method allows for easier migration of measurement chains when
internal module layout changes (hopefully it will not). As the json
files are just text the user can just replace the module name with
the new one if required.