Fix V display update when Vread is not supplied in the callback
Function `updateHistory` can be called with or without a supplied Vread.
If Vread was not supplied that defaulted to `g.Vread` (global
read-out voltage). However the default parameter assignment is taken
at the moment the function is instantiated. So it was impossible to
implicitly propagate the update Vread value to `updateHistory`. That
now defaults to None and the value is read the moment the function is
called.

This is a backport from the pypify branch