Add a deferred update tree mechanism
History tree is typically populated every time the updateTree signal
is called. When operating on single devices and self-contained modules
this mechanism works well. However in modules that operate across
multiple devices or in modules that do a multitude of distinct
operations (for instance 80 reads) this will lead to an enormous amount
of tree updates. This commit adds an option to defer tree updates until
the underlying operation has finished. This is exposed as an additional
argument to `BaseProgPanel.execute`. When the `deferredUpdate` argument
is True tree update will be done only after the module thread has
been released.