Use `np.nan` instead of None for empty markers
Newer versions of pyqtgraph check whether numpy.nan is present in an
array to calculate bounding rectangles. This will fail if a non
numpy-compatible value is present, such as None that was used for
values that are not supposed to be immediately displayed. This commit
alters that to use numpy.nan which should have been the original
approach anyway.